Promote the "evolve" operation to the base class
Invoking the model using data from the previous timestep is not
a UKF-specific task
